fix(math): fixed some conversion warns

+++ portal2-internal/util/math/math.cpp
@@ -22,7 +22,7 @@ namespace util::math {
 	}
 
 	void vector_angles( vec3_t& forward, ang_t& angles ) {
-		float tmp, yaw, pitch;
+		double tmp, yaw, pitch;
 
 		if ( forward.y == 0 && forward.x == 0 ) {
 			yaw = 0;
@@ -32,19 +32,19 @@ namespace util::math {
 				pitch = 90;
 		}
 		else {
-			yaw = ( math::atan2( forward.y, forward.x ) * 180 / pi );
-			if ( yaw < 0 )
-				yaw += 360;
+			yaw = ( math::atan2( forward.y, forward.x ) * 180.0 / pi );
+			if ( yaw < 0.0 )
+				yaw += 360.0;
 
 			tmp = math::sqrt( forward.x * forward.x + forward.y * forward.y );
-			pitch = ( math::atan2( -forward.z, tmp ) * 180 / pi );
-			if ( pitch < 0 )
-				pitch += 360;
+			pitch = ( math::atan2( -forward.z, tmp ) * 180.0 / pi );
+			if ( pitch < 0.0 )
+				pitch += 360.0;
 		}
 
-		angles.x = pitch;
-		angles.y = yaw;
-		angles.z = 0;
+		angles.x = static_cast< float >( pitch );
+		angles.y = static_cast< float >( yaw );
+		angles.z = 0.f;
 	}
 
 	void movement_fix( float& forwardmove, float& sidemove, ang_t& angles, ang_t wish_angles ) {
@@ -55,7 +55,7 @@ namespace util::math {
 		auto speed = math::sqrt( sidemove * sidemove + forwardmove * forwardmove );
 		auto yaw = math::atan2( sidemove, forwardmove ) + delta;
 
-		forwardmove = math::cos( yaw ) * speed;
-		sidemove = math::sin( yaw ) * speed;
+		forwardmove = static_cast< float >( math::cos( yaw ) * speed );
+		sidemove = static_cast< float >( math::sin( yaw ) * speed );
 	}
 }
